+++ sim-operations.h
@@ -375,6 +375,7 @@ typedef enum {
 #define save_allele_counts         gsc_save_allele_counts
 #define save_pedigrees             gsc_save_pedigrees
 #define save_bvs                   gsc_save_bvs
+#define save_local_bvs             gsc_save_local_bvs
 
 #endif
 
@@ -1757,7 +1758,7 @@ gsc_MarkerBlocks gsc_load_blocks(const gsc_SimData* d, const char* block_file);
 gsc_DecimalMatrix gsc_calculate_local_bvs(const gsc_SimData* d, 
                                                 const gsc_GroupNum group,
                                                 const gsc_MarkerBlocks b, 
-                                                const gsc_EffectID effID);
+                                                const gsc_EffectID effID); 
 void gsc_calculate_optimal_haplotype(const gsc_SimData* d, 
                                       const gsc_EffectID effID, 
                                       const char symbol_na,
@@ -1819,6 +1820,12 @@ void gsc_save_allele_counts(const char* fname,
 void gsc_save_pedigrees(const char* fname, const gsc_SimData* d, const gsc_GroupNum groupID, 
                         const _Bool full_pedigree);
 void gsc_save_bvs(const char* fname, const gsc_SimData* d, const gsc_GroupNum groupID, const gsc_EffectID effID);
+void gsc_save_local_bvs(const char* fname, 
+						const gsc_SimData* d, 
+						const gsc_GroupNum groupID, 
+						const gsc_MarkerBlocks b,
+						const gsc_EffectID effID, 
+						const _Bool headers);
 
 // Utility and helper saving functions
 void gsc_save_utility_markerblocks(FILE* f, const gsc_MarkerBlocks b, const GSC_GENOLEN_T n_markers, 
@@ -1834,6 +1841,8 @@ void gsc_save_utility_allele_counts(FILE* f,
 void gsc_save_utility_pedigrees(FILE* f, gsc_BidirectionalIterator* targets,
         const _Bool full_pedigree, const gsc_AlleleMatrix* parent_pedigree_store);
 void gsc_save_utility_bvs(FILE* f, gsc_BidirectionalIterator* targets, const gsc_MarkerEffects* eff);
+
+void gsc_save_utility_dmatrix(FILE* f, DecimalMatrix* dec, char** row_headers, char** col_headers, _Bool dim1_is_columns);
 
 // static GSC_LOGICVAL gsc_helper_is_marker_in_chr(const GSC_GENOLEN_T markerix, const gsc_LinkageGroup chr, double* pos);