+++ NEWS.md
@@ -11,7 +11,7 @@ Latest News       {#news}
 ## Improvements
 
 - Generalisation of header-parsing logic of `load_mapfile` and `load_effectfile`, to prepare for accepting further columns in future.
-- BREAKING CHANGE: `calculate_local_bvs` now returns local BVs in a vector and can be called with a group number or NO_GROUP, like other `calculate_` functions. The ability to print a matrix of local BVs to a file will be re-added soon. Removed `calculate_group_local_bvs` as its function was subsumed.
+- BREAKING CHANGE: `calculate_local_bvs` now returns local BVs in a DecimalMatrix and can be called with a group number or NO_GROUP, like other `calculate_` functions. Removed `calculate_group_local_bvs` as its function was subsumed. To calculate and save local BVs to a file, use the function `save_local_bvs`.
 - `EffectMatrix` has been replaced with new struct `MarkerEffects`. To match, `gsc_delete_effect_matrix` has been renamed to `gsc_delete_effect_set`.
 - Makefile has a new target "sharedlib" (run `make sharedlib`) to compile genomicSimulation as a shared library, rather than a standalone executable that runs the test suite.
 - The conversion script that creates the library for the genomicSimulation R package now converts `fprintf(stderr, "msg")` to `Rprintf("NOTE! msg")`, instead of `warning("msg")`, so that non-stopping errors like these are still printed if a segfault occurs. (By default, R hides `warning` messages until the end of execution scope, and so loses them if a segfault occurs).